+++ main.py
@@ -11,6 +11,7 @@ def commands():
                         type=str)
 
     parser.add_argument("--bms", help="Get battery BMS info", action="store_true")
+    parser.add_argument("-t", "--timeout", help="Bluetooth response timeout in seconds (default: 4)", type=int, default=4)
     parser.add_argument("--pair", help="Pair with device before interacting", action="store_true")
     parser.add_argument("-s", "--services", help="List device GATT services and characteristics", action="store_true")
     parser.add_argument("--verbose", help="Verbose logs", action="store_true")
@@ -31,7 +32,7 @@ def main():
         logger.setLevel(logging.DEBUG)
         logger.addHandler(handler)
 
-    battery = BatteryInfo(args.DEVICE_MAC, args.pair, logger)
+    battery = BatteryInfo(args.DEVICE_MAC, args.pair, args.timeout, logger)
 
     if args.services:
         request = battery.get_request()